requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
Django==3.0.2
djangorestframework==3.11.0
idna==2.8
inflection==0.3.1
joblib==0.14.1
more-itertools==8.1.0
numpy==1.18.1
pandas==0.25.3
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
six==1.13.0
sqlparse==0.3.0
urllib3==1.25.7
